✅ SAEVI | Medium | Java | Runtime: N/A | Memory: N/A | 2026-09-07

+++ codechef/medium/SAEVI/solution.java
@@ -8,14 +8,14 @@ class Codechef
 	{
 		Scanner a=new Scanner(System.in);
 		int n=a.nextInt();
-		int k=a.nextInt();
-		int g=2*k;
-		int[] arr=new int[n];
+		long k=a.nextLong();
+		long g=2*k;
+		long[] arr=new long[n];
 		for(int i=0;i<n;i++){
 		    arr[i]=a.nextInt();
 		    
 		}
-		int s=0;
+		long s=0;
 		for(int i=0;i<n;i+=2){
 		    if(i==0||i%2==0){
 		        if(arr[i]>g)
